+++ ai-terminal/src-tauri/src/command/autocomplete/mod.rs
@@ -1,0 +1,1 @@
+pub mod autocomplete_command;
