+++ ai-terminal/src-tauri/src/command/mod.rs
@@ -1,3 +1,4 @@
+pub mod autocomplete;
 pub mod core;
 pub mod git_commands;
 pub mod types;
