+++ ai-terminal/src-tauri/src/ollama/mod.rs
@@ -1,1 +1,2 @@
+pub mod model_request;
 pub mod types;
