+++ ai-terminal/src-tauri/src/ollama/model_request/mod.rs
@@ -1,0 +1,1 @@
+pub mod request;
