+++ ai-terminal/src-tauri/src/utils/mod.rs
@@ -1,2 +1,3 @@
+pub mod command;
 pub mod file_system_utils;
 pub mod operating_system_utils;
