+++ ai-terminal/src-tauri/src/utils/operating_system_utils.rs
@@ -1,3 +1,6 @@
+use crate::command::types::command_manager::CommandManager;
+use tauri::State;
+
 // Add a helper function to get the OS information
 pub fn get_operating_system() -> String {
     #[cfg(target_os = "windows")]
@@ -17,4 +20,19 @@ pub fn get_operating_system() -> String {
 pub fn get_system_environment_variables() -> Result<Vec<(String, String)>, String> {
     let env_vars: Vec<(String, String)> = std::env::vars().collect();
     Ok(env_vars)
+}
+
+#[tauri::command]
+pub fn get_current_pid(
+    session_id: String,
+    command_manager: State<'_, CommandManager>,
+) -> Result<u32, String> {
+    let states = command_manager.commands.lock().map_err(|e| e.to_string())?;
+    let key = session_id;
+
+    if let Some(state) = states.get(&key) {
+        Ok(state.pid.unwrap_or(0))
+    } else {
+        Ok(0)
+    }
 }
